add_manager_field

+++ src/Controller/ProfileController.php
@@ -40,16 +40,17 @@ class ProfileController extends AbstractController
 
             if ($imageFile) {
                 // Générez un nom de fichier unique pour l'image
-                $newFilename = uniqid().'.'.$imageFile->guessExtension();
+                $newFilename = uniqid().'.png';
 
                 // Déplacez le fichier dans le répertoire 'public/images'
                 $imageFile->move(
                     $this->getParameter('images_directory'),
                     $newFilename
                 );
+                $profile->setPassport($newFilename);
 
                 // Sauvegardez le chemin complet de l'image dans l'entité Profile
-                $profile->setPassport($this->getParameter('images_directory').'/'.$newFilename);
+              //  $profile->setPassport($this->getParameter('images_directory').'/'.$newFilename);
             }
 
             $entityManager->persist($profile);
@@ -65,20 +66,42 @@ class ProfileController extends AbstractController
     /**
      * @Route("/editProfile/{id}", name="profile_edit", methods={"GET", "POST"})
      */
-    public function editProfile(Request $request,EntityManagerInterface $entityManager, $id): Response
+    public function editProfile(Request $request, EntityManagerInterface $entityManager, $id): Response
     {
         $profile = $entityManager->getRepository(Profile::class)->find($id);
+        $originalPassport = $profile->getPassport(); // Sauvegarde de l'image d'origine
         $form = $this->createForm(ProfileType::class, $profile);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $imageFile = $form->get('passport')->getData();
+
+            if ($imageFile) {
+                // Générez un nom de fichier unique pour la nouvelle image
+                $newFilename = uniqid() . '.png';
+
+                // Déplacez le fichier dans le répertoire 'public/images'
+                $imageFile->move(
+                    $this->getParameter('images_directory'),
+                    $newFilename
+                );
+
+                // Mettez à jour le champ "passport" avec le nouveau nom de fichier
+                $profile->setPassport($newFilename);
+            } else {
+                // Si aucun fichier n'a été téléchargé, conservez l'image d'origine
+                $profile->setPassport($originalPassport);
+            }
+
             $entityManager->flush();
             return $this->redirectToRoute('profile_index');
         }
+
         return $this->render('profile/edit_profile.html.twig', [
             'form' => $form->createView(),
         ]);
     }
+
     /**
      * @Route("/deleteProfile/{id}", name="profile_delete")
      */
@@ -90,10 +113,17 @@ class ProfileController extends AbstractController
             throw $this->createNotFoundException('Profil non trouvé.');
         }
 
-        $entityManager->remove($profile);
-        $entityManager->flush();
+        // Vérifiez si le paramètre "delete" est présent dans la requête
+        if ($request->query->get('delete')) {
+            // Si oui, supprimez le profil
+            $entityManager->remove($profile);
+            $entityManager->flush();
+            return $this->redirectToRoute('profile_index');
+        }
 
-        return $this->redirectToRoute('profile_index', [], Response::HTTP_SEE_OTHER);
+        return $this->render('Profile/profile_delete.html.twig', [
+            'profile' => $profile,
+        ]);
     }
